FROMGIT: media: venus: keep resolution when adjusting format
When checking a format for validity, the resolution is reset to 1280x720
whenever the pixel format is not supported. This behavior can mislead
user-space into believing that this is the only resolution supported,
and looks strange considering that if we try/set the same format with
just the pixel format changed to a valid one, the call will this time
succeed without altering the resolution.

Resolution is managed independently of the pixel format, so remove this
reset.

Signed-off-by: Alexandre Courbot <acourbot@chromium.org>
Acked-by: Stanimir Varbanov <stanimir.varbanov@linaro.org>
Signed-off-by: Hans Verkuil <hans.verkuil@cisco.com>
Signed-off-by: Mauro Carvalho Chehab <mchehab+samsung@kernel.org>

BUG=b:79450911
TEST=Checked that Chrome could properly allocate buffers.

(cherry picked from commit 1946117b8f135a62eea5cfa18be63b1741174b9f git://linuxtv.org/media_tree.git master)

Change-Id: Ic5e7973c35b13f3a032000addd677398c7fe8272
Reviewed-on: https://chromium-review.googlesource.com/1158086
Commit-Ready: Alexandre Courbot <acourbot@chromium.org>
Tested-by: Alexandre Courbot <acourbot@chromium.org>
Reviewed-by: Tomasz Figa <tfiga@chromium.org>