dt-bindings: opp: Introduce opp-level bindings
Add opp-level as an additional property in the OPP node to describe
the performance level of the device.

On some SoCs (especially from Qualcomm and MediaTek) this value
is communicated to a remote microprocessor by the CPU, which
then takes some actions (like adjusting voltage values across various
rails) based on the value passed.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>