WIP: make genpd_dev_pm_attach_by_id/name leave the power domain on
genpd_dev_pm_attach() leaves the genpd ON when its attached to a
device, however when there are multiple genpds per device and
genpd_dev_pm_attach_by_id/name() is used instead to attach multiple
genpds to virtual devices, they are left disabled.

UFS for instance seems to rely on the  genpd to be ON when
ufshcd_pltfrm_init() is invoked.

Need to dig into why this change in behaviour with one genpd vs
multiple genpd but looks like there is no wasy way to handle this
in UFS if the genpd is left disabled.

Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>