WIP: remoteproc: qcom: mss: Explicitly set MSS CE SIDs to S2CR bypass
The modem sub-system on SC7180 has direct access to DDR through memnoc.
However there exists a second path to DDR which the MSS CE (crypto
engine sub-component of MSS) uses. This route has a SMMU in its path.
SMMU is not expected to provide access control/translation for these SIDs
since sandboxing of the modem is achieved through hyp assign SMC calls.
Hence these SIDs are expected to be explicity set to S2CR bypass. In the
absence of TZ programmming these SIDs into the SMMU and marking the SMRs
as secure, on SC7180 these get over-written during SMMU initialitation and
result in the global fault during modem boot-up.

This is work in progress, fix this for now by introducing an optional
"crypto-engine" sub-child to MSS and having its SIDs explicity set to
S2CR bypass. The final fix may still be done in the secure firmware.

Change-Id: Ifbe907f709b4fb4029ec429e24082a5c780b6aee
Signed-off-by: Sibi Sankar <sibis@codeaurora.org>