mmc: sdhci-msm: Use OPP API to set clk/perf state
On some qualcomm SoCs we need to vote on a performance state of a power
domain depending on the clock rates. Hence move to using OPP api to set
the clock rate and performance state specified in the OPP table.
On platforms without an OPP table, dev_pm_opp_set_rate() is eqvivalent to
clk_set_rate()

Change-Id: I51e8779a637f9a52bc01fa458b0b49ae6115ac0e
Signed-off-by: Rajendra Nayak <rnayak@codeaurora.org>