FROMLIST: irqdomain: add bus token DOMAIN_BUS_WAKEUP
A single controller can handle normal interrupts and wake-up interrupts
independently, with a different numbering space. It is thus crucial to
allow the driver for such a controller discriminate between the two.

A simple way to do so is to tag the wake-up irqdomain with a "bus token"
that indicates the wake-up domain. This slightly abuses the notion of
bus, but also radically simplifies the design of such a driver. Between
two evils, we choose the least damaging.

Suggested-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Lina Iyer <ilina@codeaurora.org>
(am from https://patchwork.kernel.org/patch/11145365/)
(also found at https://lkml.kernel.org/r/1568411962-1022-2-git-send-email-ilina@codeaurora.org)

BUG=none
TEST=Build and boot

Signed-off-by: Maulik Shah <mkshah@codeaurora.org>