FROMGIT: soc: qcom: llcc: Move regmap config to local variable
This is now a global variable that we're modifying to fix the name.
That isn't terribly thread safe and it's not necessary to be a global so
let's just move this to a local variable instead. This saves space in
the symtab and actually reduces kernel image size because the regmap
config is large and we can replace the initialization of that structure
with a memset and a few member assignments.

Cc: Venkata Narendra Kumar Gutta <vnkgutta@codeaurora.org>
Reviewed-by: Evan Green <evgreen@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
(cherry picked from commit acbf73bfa02881ba9e0532ba1a5f5beec573af9f
 git://git.kernel.org/pub/scm/linux/kernel/git/qcom/linux.git drivers-for-5.5)

BUG=None
TEST=None

Change-Id: Icbdc547e2f512a125a406e5f1926fabd1d52d62f
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>