CHROMIUM: config: Turn off MSM8916 / A53 stuff in Qualcomm config
This is:
  cat <<EOF | tee -a \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config

CONFIG_QCOM_A53PLL=n
CONFIG_QCOM_CLK_APCS_MSM8916=n

EOF
  ./chromeos/scripts/kernelconfig olddefconfig
...atop an already normalized config.

We don't support those MSM8916 or any SoCs with A53 in this config.
If we were going to support them we'd probaly want to enable more
stuff, including (possibly) some extra errata.  Turn the configs
off so there's no illusion of support.

NOTE: leave the configs on in the generic config.  I have no idea
if boards with those SoCs would boot with the generic config, but
in general we want to compile "lots of stuff" with the generic
config.

BUG=b:143638832
TEST=boot bubs

Change-Id: I36292c8dde5f4bb494eeaec57827404c78d09ea0
Signed-off-by: Douglas Anderson <dianders@chromium.org>